+++ goyave.go
@@ -19,15 +19,19 @@ import (
 )
 
 var (
-	server         *http.Server
-	redirectServer *http.Server
-	sigChannel     chan os.Signal
-	stopChannel    chan bool
-	hookChannel    chan bool
+	server             *http.Server
+	redirectServer     *http.Server
+	router             *Router
+	maintenanceHandler http.Handler
+	sigChannel         chan os.Signal
+	stopChannel        chan bool
+	hookChannel        chan bool
 
-	startupHooks []func()
-	ready        bool = false
-	mutex             = &sync.RWMutex{}
+	startupHooks       []func()
+	ready              bool = false
+	maintenanceEnabled bool = false
+	mutex                   = &sync.RWMutex{}
+	once               sync.Once
 )
 
 // IsReady returns true if the server has finished initializing and
@@ -83,9 +87,41 @@ func Start(routeRegistrer func(*Router)) {
 		database.Migrate()
 	}
 
-	router := newRouter()
+	router = newRouter()
 	routeRegistrer(router)
 	startServer(router)
+}
+
+// EnableMaintenance replace the main server handler with the "Service Unavailable" handler.
+func EnableMaintenance() {
+	mutex.Lock()
+	server.Handler = getMaintenanceHandler()
+	maintenanceEnabled = true
+	mutex.Unlock()
+}
+
+// DisableMaintenance replace the main server handler with the original router.
+func DisableMaintenance() {
+	mutex.Lock()
+	server.Handler = router.muxRouter
+	maintenanceEnabled = false
+	mutex.Unlock()
+}
+
+// IsMaintenanceEnabled return true if the server is currently in maintenance mode.
+func IsMaintenanceEnabled() bool {
+	mutex.RLock()
+	defer mutex.RUnlock()
+	return maintenanceEnabled
+}
+
+func getMaintenanceHandler() http.Handler {
+	once.Do(func() {
+		maintenanceHandler = http.HandlerFunc(func(resp http.ResponseWriter, request *http.Request) {
+			resp.WriteHeader(http.StatusServiceUnavailable)
+		})
+	})
+	return maintenanceHandler
 }
 
 // Stop gracefully shuts down the server without interrupting any
@@ -117,7 +153,9 @@ func stop(ctx context.Context) error {
 		err = server.Shutdown(ctx)
 		database.Close()
 		server = nil
+		router = nil
 		ready = false
+		maintenanceEnabled = false
 		if redirectServer != nil {
 			redirectServer.Shutdown(ctx)
 			<-stopChannel
@@ -193,6 +231,11 @@ func startServer(router *Router) {
 		Handler:      router.muxRouter,
 	}
 
+	if config.GetBool("maintenance") {
+		server.Handler = getMaintenanceHandler()
+		maintenanceEnabled = true
+	}
+
 	ln, err := net.Listen("tcp", server.Addr)
 	if err != nil {
 		fmt.Println(err)
@@ -201,7 +244,6 @@ func startServer(router *Router) {
 		stop(ctx)
 		mutex.Unlock()
 		return
-		// TODO sig channel run on http error ?
 	}
 	defer ln.Close()
 	registerShutdownHook(stop)