+++ validation/validator.go
@@ -2,7 +2,6 @@ package validation
 
 import (
 	"log"
-	"net/http"
 	"reflect"
 	"strings"
 
@@ -24,65 +23,81 @@ type RuleSet map[string][]string
 // Errors is a map of validation errors with the field name as a key.
 type Errors map[string][]string
 
-var validationRules map[string]Rule = map[string]Rule{
-	"required":           validateRequired,
-	"numeric":            validateNumeric,
-	"integer":            validateInteger,
-	"min":                validateMin,
-	"max":                validateMax,
-	"between":            validateBetween,
-	"greater_than":       validateGreaterThan,
-	"greater_than_equal": validateGreaterThanEqual,
-	"lower_than":         validateLowerThan,
-	"lower_than_equal":   validateLowerThanEqual,
-	"string":             validateString,
-	"array":              validateArray,
-	"distinct":           validateDistinct,
-	"digits":             validateDigits,
-	"regex":              validateRegex,
-	"email":              validateEmail,
-	"size":               validateSize,
-	"alpha":              validateAlpha,
-	"alpha_dash":         validateAlphaDash,
-	"alpha_num":          validateAlphaNumeric,
-	"starts_with":        validateStartsWith,
-	"ends_with":          validateEndsWith,
-	"in":                 validateIn,
-	"not_in":             validateNotIn,
-	"in_array":           validateInArray,
-	"not_in_array":       validateNotInArray,
-	"timezone":           validateTimezone,
-	"ip":                 validateIP,
-	"ipv4":               validateIPv4,
-	"ipv6":               validateIPv6,
-	"json":               validateJSON,
-	"url":                validateURL,
-	"uuid":               validateUUID,
-	"bool":               validateBool,
-	"same":               validateSame,
-	"different":          validateDifferent,
-	"confirmed":          validateConfirmed,
-	"file":               validateFile,
-	"mime":               validateMIME,
-	"image":              validateImage,
-	"extension":          validateExtension,
-	"count":              validateCount,
-	"count_min":          validateCountMin,
-	"count_max":          validateCountMax,
-	"count_between":      validateCountBetween,
-	"date":               validateDate,
-	"before":             validateBefore,
-	"before_equal":       validateBeforeEqual,
-	"after":              validateAfter,
-	"after_equal":        validateAfterEqual,
-	"date_equals":        validateDateEquals,
-	"date_between":       validateDateBetween,
-}
+var (
+	validationRules           map[string]Rule
+	typeDependentMessageRules []string
 
-var typeDependentMessageRules []string = []string{
-	"min", "max", "between", "size",
-	"greater_than", "greater_than_equal",
-	"lower_than", "lower_than_equal",
+	// Rules that check the data type and can be used to validate arrays.
+	typeRules []string
+)
+
+func init() {
+	validationRules = map[string]Rule{
+		"required":           validateRequired,
+		"numeric":            validateNumeric,
+		"integer":            validateInteger,
+		"min":                validateMin,
+		"max":                validateMax,
+		"between":            validateBetween,
+		"greater_than":       validateGreaterThan,
+		"greater_than_equal": validateGreaterThanEqual,
+		"lower_than":         validateLowerThan,
+		"lower_than_equal":   validateLowerThanEqual,
+		"string":             validateString,
+		"array":              validateArray,
+		"distinct":           validateDistinct,
+		"digits":             validateDigits,
+		"regex":              validateRegex,
+		"email":              validateEmail,
+		"size":               validateSize,
+		"alpha":              validateAlpha,
+		"alpha_dash":         validateAlphaDash,
+		"alpha_num":          validateAlphaNumeric,
+		"starts_with":        validateStartsWith,
+		"ends_with":          validateEndsWith,
+		"in":                 validateIn,
+		"not_in":             validateNotIn,
+		"in_array":           validateInArray,
+		"not_in_array":       validateNotInArray,
+		"timezone":           validateTimezone,
+		"ip":                 validateIP,
+		"ipv4":               validateIPv4,
+		"ipv6":               validateIPv6,
+		"json":               validateJSON,
+		"url":                validateURL,
+		"uuid":               validateUUID,
+		"bool":               validateBool,
+		"same":               validateSame,
+		"different":          validateDifferent,
+		"confirmed":          validateConfirmed,
+		"file":               validateFile,
+		"mime":               validateMIME,
+		"image":              validateImage,
+		"extension":          validateExtension,
+		"count":              validateCount,
+		"count_min":          validateCountMin,
+		"count_max":          validateCountMax,
+		"count_between":      validateCountBetween,
+		"date":               validateDate,
+		"before":             validateBefore,
+		"before_equal":       validateBeforeEqual,
+		"after":              validateAfter,
+		"after_equal":        validateAfterEqual,
+		"date_equals":        validateDateEquals,
+		"date_between":       validateDateBetween,
+	}
+
+	typeDependentMessageRules = []string{
+		"min", "max", "between", "size",
+		"greater_than", "greater_than_equal",
+		"lower_than", "lower_than_equal",
+	}
+
+	typeRules = []string{
+		"numeric", "integer", "timezone", "ip",
+		"ipv4", "ipv6", "json", "url", "uuid",
+		"bool", "date", "string",
+	}
 }
 
 // AddRule register a validation rule.
@@ -103,25 +118,26 @@ func AddRule(name string, typeDependentMessage bool, rule Rule) {
 	}
 }
 
-// Validate the given request with the given rule set
-// If all validation rules pass, returns nil
-func Validate(request *http.Request, data map[string]interface{}, rules RuleSet, language string) Errors {
+// Validate the given data with the given rule set.
+// If all validation rules pass, returns an empty "validation.Errors".
+// Third parameter tells the function if the data comes from a JSON request.
+// Last parameter sets the language of the validation error messages.
+func Validate(data map[string]interface{}, rules RuleSet, isJSON bool, language string) Errors {
 	var malformedMessage string
-	if request.Header.Get("Content-Type") == "application/json" {
-		malformedMessage = "Malformed JSON"
+	if isJSON {
+		malformedMessage = lang.Get(language, "malformed-json")
 	} else {
-		malformedMessage = "Malformed request"
+		malformedMessage = lang.Get(language, "malformed-request")
 	}
 	if data == nil {
 		return map[string][]string{"error": {malformedMessage}}
 	}
 
-	return validate(request, data, rules, language)
+	return validate(data, isJSON, rules, language)
 }
 
-func validate(request *http.Request, data map[string]interface{}, rules RuleSet, language string) Errors {
+func validate(data map[string]interface{}, isJSON bool, rules RuleSet, language string) Errors {
 	errors := Errors{}
-	isJSON := request.Header.Get("Content-Type") == "application/json"
 
 	for fieldName, field := range rules {
 		if !isNullable(field) && data[fieldName] == nil {
@@ -141,14 +157,64 @@ func validate(request *http.Request, data map[string]interface{}, rules RuleSet,
 				}
 				continue
 			}
-			ruleName, params := parseRule(rule)
-			if !validationRules[ruleName](fieldName, data[fieldName], params, data) {
-				message := processPlaceholders(fieldName, ruleName, params, getMessage(ruleName, data[fieldName], language), language)
-				errors[fieldName] = append(errors[fieldName], message)
+			ruleName, arrayDimensions, params := parseRule(rule)
+
+			if arrayDimensions > 0 {
+				if ok, errorValue := validateRuleInArray(ruleName, fieldName, arrayDimensions, data, params); !ok {
+					errors[fieldName] = append(
+						errors[fieldName],
+						processPlaceholders(fieldName, ruleName, params, getMessage(ruleName, *errorValue, language, arrayDimensions), language),
+					)
+				}
+			} else if !validationRules[ruleName](fieldName, data[fieldName], params, data) {
+				errors[fieldName] = append(
+					errors[fieldName],
+					processPlaceholders(fieldName, ruleName, params, getMessage(ruleName, reflect.ValueOf(data[fieldName]), language, arrayDimensions), language),
+				)
 			}
 		}
 	}
 	return errors
+}
+
+func validateRuleInArray(ruleName, fieldName string, arrayDimensions uint8, data map[string]interface{}, params []string) (bool, *reflect.Value) {
+	if t := GetFieldType(data[fieldName]); t != "array" {
+		log.Panicf("Cannot validate array values on non-array field %s of type %s", fieldName, t)
+	}
+
+	converted := false
+	var convertedArr reflect.Value
+	list := reflect.ValueOf(data[fieldName])
+	length := list.Len()
+	for i := 0; i < length; i++ {
+		v := list.Index(i)
+		value := v.Interface()
+		tmpData := map[string]interface{}{fieldName: value}
+		if arrayDimensions > 1 {
+			ok, errorValue := validateRuleInArray(ruleName, fieldName, arrayDimensions-1, tmpData, params)
+			if !ok {
+				return false, errorValue
+			}
+		} else if !validationRules[ruleName](fieldName, value, params, tmpData) {
+			return false, &v
+		}
+
+		// Update original array if value has been modified.
+		if ruleName == "array" {
+			if !converted { // Ensure field is a two dimensional array of the correct type
+				convertedArr = reflect.MakeSlice(reflect.SliceOf(reflect.TypeOf(tmpData[fieldName])), 0, length)
+				converted = true
+			}
+			convertedArr = reflect.Append(convertedArr, reflect.ValueOf(tmpData[fieldName]))
+		} else {
+			v.Set(reflect.ValueOf(tmpData[fieldName]))
+		}
+	}
+
+	if converted {
+		data[fieldName] = convertedArr.Interface()
+	}
+	return true, nil
 }
 
 func convertArray(isJSON bool, fieldName string, field []string, data map[string]interface{}) {
@@ -165,11 +231,16 @@ func convertArray(isJSON bool, fieldName string, field []string, data map[string
 	}
 }
 
-func getMessage(rule string, value interface{}, language string) string {
+func getMessage(rule string, value reflect.Value, language string, arrayDimensions uint8) string {
 	langEntry := "validation.rules." + rule
 	if isTypeDependent(rule) {
-		langEntry = langEntry + "." + GetFieldType(value)
+		langEntry += "." + getFieldType(value)
 	}
+
+	if arrayDimensions > 0 {
+		langEntry += ".array"
+	}
+
 	return lang.Get(language, langEntry)
 }
 
@@ -182,15 +253,18 @@ func getMessage(rule string, value interface{}, language string) string {
 // - "file" if the value is a slice of "filesystem.File"
 // - "unsupported" otherwise
 func GetFieldType(value interface{}) string {
-	rv := reflect.ValueOf(value)
-	kind := rv.Kind().String()
+	return getFieldType(reflect.ValueOf(value))
+}
+
+func getFieldType(value reflect.Value) string {
+	kind := value.Kind().String()
 	switch {
 	case strings.HasPrefix(kind, "int"), strings.HasPrefix(kind, "uint") && kind != "uintptr", strings.HasPrefix(kind, "float"):
 		return "numeric"
 	case kind == "string":
 		return "string"
 	case kind == "slice":
-		if rv.Type().String() == "[]filesystem.File" {
+		if value.Type().String() == "[]filesystem.File" {
 			return "file"
 		}
 		return "array"
@@ -201,6 +275,10 @@ func GetFieldType(value interface{}) string {
 
 func isTypeDependent(rule string) bool {
 	return helper.Contains(typeDependentMessageRules, rule)
+}
+
+func isArrayType(rule string) bool {
+	return helper.Contains(typeRules, rule)
 }
 
 func isRequired(field []string) bool {
@@ -215,9 +293,10 @@ func isArray(field []string) bool {
 	return helper.Contains(field, "array")
 }
 
-func parseRule(rule string) (string, []string) {
+func parseRule(rule string) (string, uint8, []string) {
 	indexName := strings.Index(rule, ":")
 	params := []string{}
+	validatesArray := uint8(0)
 	var ruleName string
 	if indexName == -1 {
 		if strings.Count(rule, ",") > 0 {
@@ -229,11 +308,25 @@ func parseRule(rule string) (string, []string) {
 		params = strings.Split(rule[indexName+1:], ",") // TODO how to escape comma?
 	}
 
+	if ruleName[0] == '>' {
+		for ruleName[0] == '>' {
+			ruleName = ruleName[1:]
+			validatesArray++
+		}
+
+		switch ruleName {
+		case "confirmed", "file", "mime", "image", "extension", "count",
+			"count_min", "count_max", "count_between":
+			log.Panicf("Cannot use rule \"%s\" in array validation", ruleName)
+		}
+
+	}
+
 	if _, exists := validationRules[ruleName]; !exists {
 		log.Panicf("Rule \"%s\" doesn't exist", ruleName)
 	}
 
-	return ruleName, params
+	return ruleName, validatesArray, params
 }
 
 // RequireParametersCount checks if the given parameters slice has at least "count" elements.