+++ validation/validator_test.go
@@ -1,8 +1,7 @@
 package validation
 
 import (
-	"net/http/httptest"
-	"strings"
+	"reflect"
 	"testing"
 
 	"github.com/System-Glitch/goyave/v2/lang"
@@ -37,12 +36,17 @@ func (suite *ValidatorTestSuite) TestIsArray() {
 	suite.False(isArray([]string{"string", "min:5", "required"}))
 }
 
+func (suite *ValidatorTestSuite) TestArrayType() {
+	suite.True(isArrayType("integer"))
+	suite.False(isArrayType("file"))
+}
+
 func (suite *ValidatorTestSuite) TestParseRule() {
-	rule, params := parseRule("required")
+	rule, _, params := parseRule("required")
 	suite.Equal("required", rule)
 	suite.Equal(0, len(params))
 
-	rule, params = parseRule("min:5")
+	rule, _, params = parseRule("min:5")
 	suite.Equal("min", rule)
 	suite.Equal(1, len(params))
 	suite.Equal("5", params[0])
@@ -54,11 +58,28 @@ func (suite *ValidatorTestSuite) TestParseRule() {
 	suite.Panics(func() {
 		parseRule("invalidrule")
 	})
+
+	rule, validatesArray, params := parseRule(">min:3")
+	suite.Equal("min", rule)
+	suite.Equal(1, len(params))
+	suite.Equal("3", params[0])
+	suite.Equal(uint8(1), validatesArray)
+
+	suite.Panics(func() {
+		parseRule(">file")
+	})
+
+	rule, validatesArray, params = parseRule(">>max:5")
+	suite.Equal("max", rule)
+	suite.Equal(1, len(params))
+	suite.Equal("5", params[0])
+	suite.Equal(uint8(2), validatesArray)
 }
 
 func (suite *ValidatorTestSuite) TestGetMessage() {
-	suite.Equal("The :field is required.", getMessage("required", "test", "en-US"))
-	suite.Equal("The :field must be at least :min.", getMessage("min", 42, "en-US"))
+	suite.Equal("The :field is required.", getMessage("required", reflect.ValueOf("test"), "en-US", 0))
+	suite.Equal("The :field must be at least :min.", getMessage("min", reflect.ValueOf(42), "en-US", 0))
+	suite.Equal("The :field values must be at least :min.", getMessage("min", reflect.ValueOf(42), "en-US", 1)) // TODO add all validation messages
 }
 
 func (suite *ValidatorTestSuite) TestAddRule() {
@@ -84,78 +105,204 @@ func (suite *ValidatorTestSuite) TestAddRule() {
 }
 
 func (suite *ValidatorTestSuite) TestValidate() {
-	rawRequest := httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/x-www-form-urlencoded; param=value")
-	errors := Validate(rawRequest, nil, RuleSet{}, "en-US")
+	errors := Validate(nil, RuleSet{}, false, "en-US")
 	suite.Equal(1, len(errors))
 	suite.Equal("Malformed request", errors["error"][0])
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/json")
-	errors = Validate(rawRequest, nil, RuleSet{}, "en-US")
+	errors = Validate(nil, RuleSet{}, true, "en-US")
 	suite.Equal(1, len(errors))
 	suite.Equal("Malformed JSON", errors["error"][0])
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/json")
-	errors = Validate(rawRequest, map[string]interface{}{
+	errors = Validate(map[string]interface{}{
 		"string": "hello world",
 		"number": 42,
 	}, RuleSet{
 		"string": {"required", "string"},
 		"number": {"required", "numeric", "min:10"},
-	}, "en-US")
+	}, true, "en-US")
 	suite.Equal(0, len(errors))
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/json")
 	data := map[string]interface{}{
 		"nullField": nil,
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"nullField": {"numeric"},
-	}, "en-US")
+	}, true, "en-US")
 	_, exists := data["nullField"]
 	suite.False(exists)
 	suite.Equal(0, len(errors))
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/json")
 	data = map[string]interface{}{
 		"nullField": nil,
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"nullField": {"required", "nullable", "numeric"},
-	}, "en-US")
+	}, true, "en-US")
 	_, exists = data["nullField"]
 	suite.True(exists)
 	suite.Equal(0, len(errors))
 
-	rawRequest = httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello%20world&number=42"))
-	rawRequest.Header.Set("Content-Type", "application/json")
 	data = map[string]interface{}{
 		"nullField": "test",
 	}
-	errors = Validate(rawRequest, data, RuleSet{
+	errors = Validate(data, RuleSet{
 		"nullField": {"required", "nullable", "numeric"},
-	}, "en-US")
+	}, true, "en-US")
 	_, exists = data["nullField"]
 	suite.True(exists)
 	suite.Equal(1, len(errors))
 }
 
 func (suite *ValidatorTestSuite) TestValidateWithArray() {
-	rawRequest := httptest.NewRequest("POST", "/test-route", strings.NewReader("string=hello"))
-	rawRequest.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 	data := map[string]interface{}{
 		"string": "hello",
 	}
-	errors := Validate(rawRequest, data, RuleSet{
+	errors := Validate(data, RuleSet{
 		"string": {"required", "array"},
-	}, "en-US")
+	}, false, "en-US")
 	suite.Equal("array", GetFieldType(data["string"]))
 	suite.Equal("hello", data["string"].([]string)[0])
 	suite.Equal(0, len(errors))
+}
+
+func (suite *ValidatorTestSuite) TestValidateArrayValues() {
+	data := map[string]interface{}{
+		"string": []string{"hello", "world"},
+	}
+	errors := Validate(data, RuleSet{
+		"string": {"required", "array", ">min:3"},
+	}, false, "en-US")
+	suite.Equal(0, len(errors))
+
+	data = map[string]interface{}{
+		"string": []string{"hi", ",", "there"},
+	}
+	errors = Validate(data, RuleSet{
+		"string": {"required", "array", ">min:3"},
+	}, false, "en-US")
+	suite.Equal(1, len(errors))
+
+	data = map[string]interface{}{
+		"string": []string{"johndoe@example.org", "foobar@example.org"},
+	}
+	errors = Validate(data, RuleSet{
+		"string": {"required", "array:string", ">email"},
+	}, true, "en-US")
+	suite.Equal(0, len(errors))
+
+	suite.Panics(func() {
+		validateRuleInArray("required", "string", 1, map[string]interface{}{"string": "hi"}, []string{})
+	})
+
+	// Empty array
+	data = map[string]interface{}{
+		"string": []string{},
+	}
+	errors = Validate(data, RuleSet{
+		"string": {"array", ">uuid:5"},
+	}, true, "en-US")
+	suite.Equal(0, len(errors))
+}
+
+func (suite *ValidatorTestSuite) TestValidateTwoDimensionalArray() {
+	data := map[string]interface{}{
+		"values": [][]interface{}{{"0.5", 1.42}, {0.6, 7}},
+	}
+	errors := Validate(data, RuleSet{
+		"values": {"required", "array", ">array:numeric"},
+	}, false, "en-US")
+	suite.Equal(0, len(errors))
+
+	arr, ok := data["values"].([][]float64)
+	suite.True(ok)
+	if ok {
+		suite.Equal(2, len(arr))
+		suite.Equal(0.5, arr[0][0])
+		suite.Equal(1.42, arr[0][1])
+		suite.Equal(0.6, arr[1][0])
+		suite.Equal(7.0, arr[1][1])
+	}
+
+	data = map[string]interface{}{
+		"values": [][]float64{{5, 8}, {0.6, 7}},
+	}
+	errors = Validate(data, RuleSet{
+		"values": {"required", "array", ">array:numeric", ">min:3"},
+	}, true, "en-US")
+	suite.Equal(1, len(errors))
+
+	_, ok = data["values"].([][]float64)
+	suite.True(ok)
+
+	data = map[string]interface{}{
+		"values": [][]float64{{5, 8, 6}, {0.6, 7, 9}},
+	}
+	errors = Validate(data, RuleSet{
+		"values": {"required", "array", ">array:numeric", ">min:3"},
+	}, true, "en-US")
+	suite.Equal(0, len(errors))
+
+	data = map[string]interface{}{
+		"values": [][]float64{{5, 8}, {3, 7}},
+	}
+	errors = Validate(data, RuleSet{
+		"values": {"required", "array", ">array:numeric", ">>min:3"},
+	}, true, "en-US")
+	suite.Equal(0, len(errors))
+
+	data = map[string]interface{}{
+		"values": [][]float64{{5, 8}, {0.6, 7}},
+	}
+	errors = Validate(data, RuleSet{
+		"values": {"required", "array", ">array:numeric", ">>min:3"},
+	}, true, "en-US")
+	suite.Equal(1, len(errors))
+}
+
+func (suite *ValidatorTestSuite) TestValidateNDimensionalArray() {
+	data := map[string]interface{}{
+		"values": [][][]interface{}{
+			{{"0.5", 1.42}, {0.6, 4, 3}},
+			{{"0.6", "1.43"}, {}, {2}},
+		},
+	}
+	errors := Validate(data, RuleSet{
+		"values": {"required", "array", ">array", ">>array:numeric", ">max:3", ">>>max:4"},
+	}, true, "en-US")
+	suite.Equal(0, len(errors))
+
+	arr, ok := data["values"].([][][]float64)
+	suite.True(ok)
+	if ok {
+		suite.Equal(2, len(arr))
+		suite.Equal(2, len(arr[0]))
+		suite.Equal(3, len(arr[1]))
+		suite.Equal(0.5, arr[0][0][0])
+		suite.Equal(1.42, arr[0][0][1])
+		suite.Equal(2.0, arr[1][2][0])
+	}
+
+	data = map[string]interface{}{
+		"values": [][][]interface{}{
+			{{"0.5", 1.42}, {0.6, 4, 3}},
+			{{"0.6", "1.43"}, {}, {2}, {4}},
+		},
+	}
+	errors = Validate(data, RuleSet{
+		"values": {"required", "array", ">array", ">>array:numeric", ">max:3", ">>>max:4"},
+	}, true, "en-US")
+	suite.Equal(1, len(errors))
+
+	data = map[string]interface{}{
+		"values": [][][]interface{}{
+			{{"0.5", 1.42}, {0.6, 9, 3}},
+			{{"0.6", "1.43"}, {}, {2}},
+		},
+	}
+	errors = Validate(data, RuleSet{
+		"values": {"required", "array", ">array", ">>array:numeric", ">max:3", ">>>max:4"},
+	}, true, "en-US")
+	suite.Equal(1, len(errors))
 }
 
 func TestValidatorTestSuite(t *testing.T) {